added ordering business logic
There is an order form in the menu show view. On that form you can select a quantity for each item. A blank quantity of course means we're not picking that item. Once the form is submitted we proceed to the Order show which contains the items you selected and lets you change the quantities. Then we should proceed to payment